+++ src/Scripts/NPCData.lua
@@ -693,7 +693,6 @@ local npcModifications = {
 						},
 						TextLineSet = {
 							BecameCloseWithThanatos01 = {
-								Name = "BecameCloseWithThanatos01",
 								GameStateRequirements = {},
 								-- Removed for context
 								-- {
@@ -878,7 +877,6 @@ local npcModifications = {
 						},
 						TextLineSet = {
 							BecameCloseWithThanatos01_B = {
-								Name = "BecameCloseWithThanatos01_B",
 								GameStateRequirements = {},
 								-- Removed for context
 								-- {
@@ -1781,7 +1779,6 @@ local npcModifications = {
 						},
 						TextLineSet = {
 							MegaeraBedroom01 = {
-								Name = "MegaeraBedroom01",
 								-- Requirements are above
 								GameStateRequirements = {},
 								{
@@ -1848,7 +1845,6 @@ local npcModifications = {
 						},
 						TextLineSet = {
 							MegaeraBedroom02 = {
-								Name = "MegaeraBedroom02",
 								-- Requirements are above
 								GameStateRequirements = {},
 								{
@@ -1940,7 +1936,6 @@ local npcModifications = {
 						},
 						TextLineSet = {
 							MegaeraBedroom02B = {
-								Name = "MegaeraBedroom02B",
 								-- Requirements are above
 								GameStateRequirements = {},
 								{
@@ -2033,7 +2028,6 @@ local npcModifications = {
 						},
 						TextLineSet = {
 							BecameCloseWithMegaera01 = {
-								Name = "BecameCloseWithMegaera01",
 								-- Requirements are above
 								GameStateRequirements = {},
 								{
@@ -2227,7 +2221,6 @@ local npcModifications = {
 						},
 						TextLineSet = {
 							BecameCloseWithMegaera01_B = {
-								Name = "BecameCloseWithMegaera01_B",
 								-- Requirements are above
 								GameStateRequirements = {},
 								{