+++ src/Scripts/NPCLogic.lua
@@ -663,16 +663,26 @@ function mod.SurpriseNPCPresentation(source, args)
 	end
 
 	AddInputBlock({ Name = "SurpriseNPCPresentation" })
-
-	game.EndAutoSprint()
-	CancelWeaponFireRequests({ Id = game.CurrentRun.Hero.ObjectId })
-	Stop({ Id = game.CurrentRun.Hero.ObjectId })
-	Halt({ Id = game.CurrentRun.Hero.ObjectId })
 	ToggleControl({ Names = { "AdvancedTooltip", }, Enabled = false })
 
 	game.killWaitUntilThreads("ReattachCameraOnInput")
 
+	-- For the Thanatos and Megaera romance scenes where the player can move before this is called
+	if args.StopHeroImmediately then
+		game.EndAutoSprint()
+		CancelWeaponFireRequests({ Id = game.CurrentRun.Hero.ObjectId })
+		Stop({ Id = game.CurrentRun.Hero.ObjectId })
+		Halt({ Id = game.CurrentRun.Hero.ObjectId })
+	end
+
 	game.wait(args.IntroWait or 0.4, game.RoomThreadName)
+
+	if not args.StopHeroImmediately then
+		game.EndAutoSprint()
+		CancelWeaponFireRequests({ Id = game.CurrentRun.Hero.ObjectId })
+		Stop({ Id = game.CurrentRun.Hero.ObjectId })
+		Halt({ Id = game.CurrentRun.Hero.ObjectId })
+	end
 
 	AngleTowardTarget({ Id = source.ObjectId, DestinationId = game.CurrentRun.Hero.ObjectId })
 	if not args.SkipPan then
@@ -684,7 +694,7 @@ function mod.SurpriseNPCPresentation(source, args)
 	RemoveInputBlock({ Name = "SurpriseNPCPresentation" })
 
 	if args.TextLineSet ~= nil then
-		game.ProcessTextLines(args.TextLineSet)
+		game.ProcessTextLines(source, args.TextLineSet)
 		mod.PlayRandomRemainingTextLines(source, args.TextLineSet)
 	end