+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Other fixes and improvements:
 - You can now salute story NPCs you meet during Zagreus' Journey runs.
 - Updated the description text of "Final Expense" offered by Orpheus to make it clearer that all shop items are always discounted.
 - If Thanatos has no new unique dialogue when he visits you at the start of a run, you can now pick up the starting boon without speaking to him first.
+- Thanatos and Megaera can now pay Sisyphus a visit.
 - Added a new dialogue background for conversations in the Temple of Styx.
 - Hades and Theseus will now react with specific voicelines for certain Gods when using Godsent Hex effects against them.
 - Hades will now react to Melinoë going dark using "Unseen Ire" in the fight against him.